working on fuzzy math library
