+++ src/Elba/Core/Components/MAT362/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(elba
 PRIVATE
-    ${CMAKE_CURRENT_LIST_DIR}/FuzzyController.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/GammaController.cpp
 PUBLIC
-    ${CMAKE_CURRENT_LIST_DIR}/FuzzyController.hpp
+    ${CMAKE_CURRENT_LIST_DIR}/GammaController.hpp
 )
