+++ src/Elba/Core/Components/MAT362/GammaController.cpp
@@ -1,0 +1,7 @@
+#include "Elba/Core/Components/MAT362/GammaController.hpp"
+
+namespace Elba
+{
+
+
+} // End of Elba namespace
